COSBETA-241: Mock domain request handling in login test helper methods
Mock the appropriate domain behaviour for each user type in the login helper methods used in most tests. Explicitly mock the domain handling in tests where the login helper methods are not used.
