+++ cosmetics-web/spec/rails_helper.rb
@@ -1,6 +1,7 @@
 # This file is copied to spec/ when you run 'rails generate rspec:install'
 require 'spec_helper'
 require 'login_helpers'
+require 'domain_helpers'
 require 'responsible_person_helpers'
 require 'file_helpers'
 require 'policy_helpers'
@@ -68,6 +69,7 @@ RSpec.configure do |config|
 
   config.include FactoryBot::Syntax::Methods
 
+  config.include DomainHelpers
   config.include LoginHelpers
   config.include ResponsiblePersonHelpers
   config.include FileHelpers
