+++ cosmetics-web/spec/system/domain_spec.rb
@@ -2,7 +2,6 @@ require 'rails_helper'
 
 RSpec.describe "Service domain", type: :system do
   after do
-    reset_domain_request_mocking
     sign_out
   end
 
@@ -46,28 +45,5 @@ RSpec.describe "Service domain", type: :system do
       assert_current_path invalid_account_path
       assert_text "Your account doesn’t allow you to use this service"
     end
-  end
-
-private
-
-  def configure_requests_for_submit_domain
-    allow_any_instance_of(DomainConstraint).to receive(:matches?).and_return(true)
-
-    allow_any_instance_of(ApplicationController).to receive(:submit_domain?).and_return(true)
-    allow_any_instance_of(ApplicationController).to receive(:search_domain?).and_return(false)
-  end
-
-  def configure_requests_for_search_domain
-    allow_any_instance_of(DomainConstraint).to receive(:matches?).and_return(true)
-
-    allow_any_instance_of(ApplicationController).to receive(:submit_domain?).and_return(false)
-    allow_any_instance_of(ApplicationController).to receive(:search_domain?).and_return(true)
-  end
-
-  def reset_domain_request_mocking
-    allow_any_instance_of(DomainConstraint).to receive(:matches?).and_call_original
-
-    allow_any_instance_of(ApplicationController).to receive(:submit_domain?).and_call_original
-    allow_any_instance_of(ApplicationController).to receive(:search_domain?).and_call_original
   end
 end
